feat: integrate AuthFlow into Login page, enhancing authentication experience and streamlining user navigation
